+++ RAnalogs/RAnEn/R/verifyRankHist.R
@@ -13,8 +13,21 @@
 
 #' RAnEn::verifyRankHist
 #' 
-#' Underdevelopment.
+#' RAnEn::verifyRankHist calculates the rank histogram.
 #' 
+#' @author Guido Cervone \email{cervone@@psu.edu}
+#' @author Martina Calovi \email{mxc895@@psu.edu}
+#' @author Laura Clemente-Harding \email{laura@@psu.edu}
+#' @author Weiming Hu \email{weiming@@psu.edu}
+#' 
+#' @param anen.ver A 4-dimensional array. This array is usually created from the `value` column of
+#' the `analogs` member in the results of \code{\link{generateAnalogs}}. The dimensions should be
+#' `[stations, times, lead times, members]`.
+#' @param obs.ver A 3-dimensional array. The dimensions should be `[stations, times, lead times]`.
+#' You can generate the array using \code{\link{alignObservations}}.
+#' @param show.progress Whether to plot a progress bar.
+#' 
+#' @md
 #' @export
 verifyRankHist <- function(anen.ver, obs.ver, show.progress = F){